Move transaction COMMIT and ROLLBACK into the finally block
A transaction must either commit or rollback.

The transaction implementation before this change depended on throwing
and catching exceptions, which prevented consumers from implementing
their own error handling in the transaction's `callback` function. That
is, if `callback` caught an error, it needed to rethrow it (or a
different error).

If an error wasn't thrown in the `callback` after an error occurred, the
transaction was erroneously committed, leaving the consuming application
with the responsibility of cleaning up bad data. This change lets us
lean on the SQLite schema and constraints to maintain data integrity.